Move `AgentRawData` to snmplib
This is obviously wrong (agent data != snmp).  The type should
most likely go the check engine, where we already have, for
example, `AgentRawDataSection`.  However, to achieve this, we
will have to clarify the layering between fetchers and
check engine otherwise, the change results in cyclic dependencies.

This is nevertheless and improvement in that snmplib and whereever
`AgentRawData` ends up are at the same level of abstraction level.
Also, we already had related types in `snmplib`.

CMK-13768
