+++ cmk/base/agent_based/discovery/_discovery.py
@@ -15,9 +15,9 @@ from cmk.utils.exceptions import OnError
 from cmk.utils.hostaddress import HostName
 from cmk.utils.labels import DiscoveredHostLabelsStore, HostLabel
 from cmk.utils.sectionname import SectionName
-from cmk.utils.type_defs import AgentRawData, Item, ServiceName
+from cmk.utils.type_defs import Item, ServiceName
 
-from cmk.snmplib.type_defs import SNMPRawData
+from cmk.snmplib.type_defs import AgentRawData, SNMPRawData
 
 from cmk.checkengine import (
     DiscoveryPlugin,
