+++ cmk/checkengine/crash_reporting.py
@@ -15,9 +15,9 @@ import cmk.utils.encoding
 import cmk.utils.paths
 from cmk.utils.hostaddress import HostName
 from cmk.utils.sectionname import SectionName
-from cmk.utils.type_defs import AgentRawData, CheckPluginNameStr, ServiceName
+from cmk.utils.type_defs import CheckPluginNameStr, ServiceName
 
-from cmk.snmplib.type_defs import SNMPBackendEnum
+from cmk.snmplib.type_defs import AgentRawData, SNMPBackendEnum
 
 from cmk.checkengine.checking import CheckPluginName
