+++ cmk/fetchers/_ipmi.py
@@ -24,7 +24,8 @@ from six import ensure_binary
 from cmk.utils.exceptions import MKFetcherError
 from cmk.utils.hostaddress import HostAddress
 from cmk.utils.log import VERBOSE
-from cmk.utils.type_defs import AgentRawData
+
+from cmk.snmplib.type_defs import AgentRawData
 
 from cmk.fetchers import Fetcher, Mode
