+++ cmk/fetchers/_nofetcher.py
@@ -11,7 +11,8 @@ from collections.abc import Mapping
 from typing import Final, NoReturn
 
 from cmk.utils.exceptions import MKFetcherError
-from cmk.utils.type_defs import AgentRawData
+
+from cmk.snmplib.type_defs import AgentRawData
 
 from cmk.fetchers import Fetcher, Mode
