+++ cmk/fetchers/snmp_backend/stored_walk.py
@@ -14,9 +14,9 @@ import cmk.utils.paths
 from cmk.utils.exceptions import MKGeneralException, MKSNMPError
 from cmk.utils.log import console
 from cmk.utils.sectionname import SectionName
-from cmk.utils.type_defs import AgentRawData
 
 from cmk.snmplib.type_defs import (
+    AgentRawData,
     OID,
     SNMPBackend,
     SNMPContextName,
