+++ tests/unit/cmk/base/agent_based/test_base_inventory.py
@@ -14,9 +14,8 @@ from cmk.utils.everythingtype import EVERYTHING
 from cmk.utils.hostaddress import HostAddress, HostName
 from cmk.utils.sectionname import SectionName
 from cmk.utils.structured_data import ImmutableTree, MutableTree, RetentionInterval, UpdateResult
-from cmk.utils.type_defs import AgentRawData
 
-from cmk.snmplib.type_defs import SNMPRawData
+from cmk.snmplib.type_defs import AgentRawData, SNMPRawData
 
 from cmk.fetchers import FetcherType
